requirements.txt
python-telegram-bot
groq
pytesseract
pillow
PyPDF2
python-docx
pandas